# docs/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cards Management System API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "localhost server"
    }
  ],
  "security": [
    {
      "Bearer Authentication": []
    }
  ],
  "tags": [
    {
      "name": "Admin Controller",
      "description": "Endpoints available only to admins"
    },
    {
      "name": "Authorization Controller",
      "description": "Endpoints for user's authorization"
    },
    {
      "name": "Profile Controller",
      "description": "Endpoint for current user's data"
    },
    {
      "name": "Cards Controller",
      "description": "Endpoints for user's cards opportunities"
    }
  ],
  "paths": {
    "/api/v1/cards/transfer": {
      "post": {
        "tags": [
          "Cards Controller"
        ],
        "summary": "Make money transfer from one card to another",
        "operationId": "makeMoneyTransfer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoneyTransferDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/cards/block-card-request": {
      "post": {
        "tags": [
          "Cards Controller"
        ],
        "summary": "Make card block request (wait for admin's confirmation)",
        "operationId": "makeCardBlockRequest",
        "parameters": [
          {
            "name": "cardId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/auth/sign-up": {
      "post": {
        "tags": [
          "Authorization Controller"
        ],
        "summary": "Reqister new account",
        "operationId": "signUp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseTokenDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/sign-in": {
      "post": {
        "tags": [
          "Authorization Controller"
        ],
        "summary": "Log into account",
        "operationId": "signIn",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseTokenDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/unblock-card": {
      "post": {
        "tags": [
          "Admin Controller"
        ],
        "summary": "Unblock user's card",
        "operationId": "unblockCard",
        "parameters": [
          {
            "name": "cardId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/admin/create-card": {
      "post": {
        "tags": [
          "Admin Controller"
        ],
        "summary": "Create new card for user",
        "operationId": "createCard",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/v1/admin/block-card": {
      "post": {
        "tags": [
          "Admin Controller"
        ],
        "summary": "Handle card block request from user",
        "operationId": "blockCard",
        "parameters": [
          {
            "name": "requestId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/users/profile": {
      "get": {
        "tags": [
          "Profile Controller"
        ],
        "summary": "Get current user's data",
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cards": {
      "get": {
        "tags": [
          "Cards Controller"
        ],
        "summary": "Get all cards for current user",
        "operationId": "getUserCards",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageCardDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cards/search": {
      "get": {
        "tags": [
          "Cards Controller"
        ],
        "summary": "Search for cards by filters",
        "operationId": "searchCards",
        "parameters": [
          {
            "name": "cardNumber",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cardHolder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "BLOCKED",
                "EXPIRED"
              ]
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageCardDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin": {
      "get": {
        "tags": [
          "Admin Controller"
        ],
        "summary": "Get all users",
        "operationId": "getAllUsers",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageUserDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/profile": {
      "get": {
        "tags": [
          "Admin Controller"
        ],
        "summary": "Get current user's data (for admins)",
        "operationId": "getCurrentUser_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/cards": {
      "get": {
        "tags": [
          "Admin Controller"
        ],
        "summary": "Get all cards",
        "operationId": "getAllCards",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageCardDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/block-requests": {
      "get": {
        "tags": [
          "Admin Controller"
        ],
        "summary": "Get all card block requests from users",
        "operationId": "getAllCardBlockRequests",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageCardBlockRequest"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/remove-user": {
      "delete": {
        "tags": [
          "Admin Controller"
        ],
        "summary": "Delete user from system",
        "operationId": "removeUser",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/admin/delete-card": {
      "delete": {
        "tags": [
          "Admin Controller"
        ],
        "summary": "Delete user's card",
        "operationId": "deleteCard",
        "parameters": [
          {
            "name": "cardId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MoneyTransferDto": {
        "type": "object",
        "properties": {
          "fromCardId": {
            "type": "string",
            "format": "uuid"
          },
          "toCardId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "fromCardId",
          "toCardId"
        ]
      },
      "SignUpRequestDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "username"
        ]
      },
      "ResponseTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "SignInRequestDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "password",
          "username"
        ]
      },
      "CardCreateDto": {
        "type": "object",
        "properties": {
          "cardNumber": {
            "type": "string"
          },
          "cardHolder": {
            "type": "string"
          },
          "expiryDate": {
            "type": "string"
          },
          "balance": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "cardHolder",
          "cardNumber",
          "expiryDate"
        ]
      },
      "CardDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "cardNumber": {
            "type": "string"
          },
          "cardHolder": {
            "type": "string"
          },
          "expiryDate": {
            "type": "string"
          },
          "cardStatus": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "BLOCKED",
              "EXPIRED"
            ]
          },
          "balance": {
            "type": "number",
            "format": "double"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "USER"
            ]
          },
          "cards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardDto"
            }
          }
        }
      },
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PageCardDto": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardDto"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "paged": {
            "type": "boolean"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
      "PageUserDto": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserDto"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "CardBlockRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "cardId": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "datetime": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "PROGRESS",
              "RESOLVED"
            ]
          }
        }
      },
      "PageCardBlockRequest": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardBlockRequest"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer Authentication": {
        "type": "http",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}